ARM: OMAP2+: add support for OMAP5 remote processors
Support has been added to enable the remoteproc device
registration for OMAP5. OMAP5, like OMAP4, also has two
processor subsystems - DSP and IPU. The IPU subsystem
though has dual Cortex-M4 processors instead of the
dual Cortex-M3 processors in OMAP4, but otherwise has
almost the same features.

The support includes adding a common function for reserving
CMA pools with the OMAP5 SoC identifier. The OMAP5 support
reuses the same data structures as that of OMAP4, and an
adjustment has been made only for DMTimers 5 and 6. These
timers are capable of PWM functionality also in OMAP5, and
the logic is adjusted (at runtime to avoid creating separate
static data structures for OMAP5) so that the timers can be
successfully acquired for the DSP sub-system.

Signed-off-by: Suman Anna <s-anna@ti.com>