OMAP5 EVM panel: LG4591: set PWM frequency
The Regulator PWM input frequency range is 1kHz - 100KHz.
With initial settings of LG4591 panel, the PWM frequency is set default
to 260Hz. Change this to 3.8KHz.

With this change the minimum brightness value is 0 and the maximum brightness
value is 255 for the OMAP5 EVM LCD panel LG4591.

Thanks to  "Larson, Scott"<s-larson2@ti.com> and "Fisher, Robb"
<rfisher@ti.com> for identifying the root cause.

Change-Id: I27c3f6838e6b8493b226ce05418e565c6beba1f3
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>