OMAPDSS: DISPC: WA: OMAP5ES1.0: Errata ID: i715
Fix the following HW issue: When the premultiply alpha is enabled and when the
global alpha is maximum, the layer is made opaque instead of blending.

Workaround: Reduce global alpha by 1 when premultiply alpha is enabled
(suggested as workaround in errata sheet) and we are blending a layer with alpha
info.

Errata ID: i715

Change-Id: Ib77342f4721203280279d1358f905ea0ffa38f18
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>
Signed-off-by: Lajos Molnar <lajos@ti.com>
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>