OMAP3+: PM: VC: setup ON voltages when setting a new operational voltage
ON voltage for LP and OFF states need to be programmed in for wakingup
at voltage we went down in instead of going up always with initial
configured for that rail. This is important as we may have varied
voltage setups when DVFS is functional. SR will be re-enabled as
part of power core code to re-adjust the resultant voltage to
optimal voltage levels.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>