ARM: OMAP5: add cm1 and cm2 save / restore support
cm1 re-uses omap4 register declarations, as the addresses and registers
are exactly the same. For cm2, a new register list is defined.

It make sense to re-name cm44xx.c to cm44xx_54xx.c as cm1 is getting re-used
and new cm2 list is added in the same file. Also modify corresponding Makefile
for common omap4/5 build.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
[girishsg@ti.com: Fix merge conflicts and re-name cm44xx to cm44xx_54xx]
Signed-off-by: Girish S G <girishsg@ti.com>