OMAPDSS: DSS: Add state notifiers before panel is disabled
Wrapping dssdev->driver->disable, so that we can notify
clients before a panel is disconnected.  Clients need to
be notified so they can cancel any operations that would
be blocked on display VSYNC events.

The notifiers can be extended to handle other state
change notifications in the future if needed by clients.

Change-Id: I8854c5460258f7293e4c8fb8daae6b7475512fd3
Signed-off-by: Lajos Molnar <lajos@ti.com>
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>