OMAPDSS: DSS: Add manager blanking support
This patch is based on the original patch-set present in
K3.0 & K3.1 kernel. This patch adds support for manager blank
based on new manager apply() changes in K3.4 DSS2 driver

Add manager->blank() to disable all overlays on a manager and
propagate it through all of the dss_cache stages.

blank() takes a single argument (besides the manager): wait_for_go.
If set, blank waits for go bit to be reset before it returns.
This ensures that the blanking has been applied to the manager.

if wait_for_go is false, blank() will still flush the callbacks
in the shadow and displayed stages of the dss cache without actually
calling apply.  One must use this mode if there are no further
VSYNCS for the display.  This is safe because the dss_cache is
updated for the blanked overlays, so the blanking will take effect
the next time configure() is called.

You want to use wait_for_go = true, before turning off a display.
This way you ensure that DISPC had a chance to turn off all overlays
on the manager.  Otherwise, DISPC may get into an undefined
state if you are trying to reuse those overlays on another manager.

Also, turning the display off while the GO bit is set may be an issue
even if the overlays are not reused on another manager.

Manager apply is skipped if display state is not active.

Just as in mgr->apply(), we have to register for the apply ISR if
the GO bit is on when trying to change manager settings, so that we
can process the blanking command when the GO bit is cleared.

Change-Id: I17d49626f696c8581f8792727a4551f2c18cede2
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>
Signed-off-by: Lajos Molnar <lajos@ti.com>
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>