OMAP5:GPIO:WA for GPIO o/p pins for OFF mode
In OMAP5 the GPIO pins 140 & 142 are configured as o/p pins
to drive WLAN and BT respectively. While attempting OFF it is
observed that these lines get pulled low resulting in loss of
connection. This is an issue in ES1.0 (OMAP5430-1.0BUG01667)
and will be fixed in ES2.0. As a workaround for ES1.0 these
pins are changed from o/p to i/p while going to OFF mode and
back as o/p pins in resume.

Added H/W BUD id in commit message and code. Also added
"id" variable in gpio_bank structure to identify the pins
from the bank which need change in direction. This will also
be removed when the WA is removed in ES 2.0

Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>