rpmsg: omx: confirm remove function is called because of a recovery
Now, we have access to the remoteproc handle for the rpdev channels
using that handle we can check the state of the remoteproc and confirm
when rpmsg_omx_remove function is called because of a recovery.

Also, this patch is adding missing protection between rpmsg_omx_remove
and rpmsg_omx_release for checking omx->state.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>