misc: emif: add support for TI DVFS framework
EMIF driver is designed for mainline kernel. Hence
it doesn't have support for TI's out-of-tree DVFS
framework. Add that support here for use in our internal
trees.

Use TI internal tree APIs to register with the framework
for being notified for voltage and frequency changes.

As part of this change, ensure that emif driver is registered
in the system as-soon as possible - postcore_initcall is used
for that reason.

Since EMIF is must have for SoC functionality, enable EMIF
by default for supported SoCs.

TI Framework will invoke voltage scale before frequency scale
at boot- this makes the error condition more or less a debug
condition. So, reduce severity of warning on call for voltage
sensitive register lower.

[nm@ti.com: beautification]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>