ARM: OMAP: USB: HOST TLL platform driver support
The platform driver for the TLL component of the OMAP USB host controller
is implemented. Depending on the TLL hardware revision, the TLL channels
are configured. The USB HS core driver uses this driver through exported
APIs from the TLL platform driver.usb_tll_enable and usb_tll_disable are
the exported APIS of the USB TLL platform driver.

The hwmod of the usb tll is retrieved and omap device build is
performed to created the platform device for the usb tll component.

The TLL specific code such as channels clocks enable/disable,
initialization functions are removed from the USBHS core
driver.

Change-Id: I15cd5423a16c53024fb516b797aba2689f340955
Signed-off-by: Keshava Munegowda <keshava_mgowda@ti.com>
Signed-off-by: Mantesh Sarashetti <mantesh@ti.com>
Reviewed-by: Partha Basak <parthab@india.ti.com>