remoteproc: avoid processing messages after a crash
Sometimes the kick message from the remoteproc is deliver in a deferred
task by the low level driver. In that scenario it is possible to receive
a message after the remoteproc has crashed. However, there is no point
processing those messages vrings can be already deleted by the recovery
process. So, we drop that kick messages

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>