ARM: OMAP5: hwmod data: add mmu device attributes for ipu and dsp
The MMU device attribute data is missing from the OMAP5 ES2 hwmod
data. This attribute data needs to be provided in order for the
proper initialization and creation of the iommu platform devices.
The lack of this data was causing a kernel null pointer data abort.

The hwmod class is also adjusted properly for the main ipu and dsp
hwmod devices to point to the omap54xx_hwmod_mmu_class. This is done
to have the OMAP iommu platform devices init properly as they are
based on 'mmu' class name.

Change-Id: I7542e68b2c8b2f75ad56f0efea5dca15e2503d3b
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>