mmc: omap_hsmmc: Errata: Fix SD card removal detection
Because of OMAP4 Silicon errata (i705), we have to turn off the
PBIAS and VMMC for SD card as soon as we get card disconnect
interrupt. Because of this, we don't wait for all higher layer
structures to be dismantled before turning off power. As a side
effect of this, we might end up getting a mmc_request
even after SD is removed and VMMC and PBIAS are turned off.
In that case, just fail the mmc_request and return immediately

Change-Id: I2d89b0040f0c050c99ba1533a9674926a1639b25
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>
Signed-off-by: Semen Protsenko <semen.protsenko@ti.com>