OMAP4+: HSI: Prevent HSI clock disable while ACWAKE is high
If HSI clocks are disabled while ACWAKE is high, it is possible that ACWAKE
goes low unexpectedly when L3INIT_PD goes to LPM.

For Link Layers which comply with MIPI, ACWAKE will go low independently of
CAWAKE, and HSI driver might cut the clocks (and Power Domain) while ACWAKE is
supposed to stay high, which causes ACWAKE to have an unpredictable value.

This is not seen on Link Layers which do not comply with MIPI, because in this
case ACWAKE and CAWAKE are made high at the same time. So thanks to CAWAKE high,
HSI clocks are not disabled while ACWAKE goes low and we don't see the issue.

Change-Id: If981c82e1d59682c0a3400f7d769575738986560
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>