OMAP: dmtimer: Fix the runtime PM API used in driver
In the current driver we use pm_runtime_put. The problem with this is
that it calls pm_request_idle which queues up runtime_idle() by a work
item in pm_wq. We see that if there is an immediate pm_runtime_get_sync
call followed by this, it returns 1 instead of 0 and there is a mismatch
in clk_enable/clk_disable calls. However if we use pm_runtime_put_sync
it calls pm_runtime_idle and and ensures synchronized clk disable. This
patch replaces pm_runtime_put with pm_runtime_put_sync in the driver.
This issue was caught by running timer test case which in a loop changes
parent clk, enables interrupt and sets up a counter and then disables it
for all the supported timers.

Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>