usb:gadget:android: set right bMaxPower config desc
'bMaxPower' field of configuration descriptor defines
maximum power that a device can draw from HOST. This
is expressed in 2-mA units when the device is operating
in high-speed mode and in 8-mA units when operating in
SuperSpeed mode. (as per USB3 spec).

Overwrite this field in the android gadget appropriately
depending on the current speed of operation.

Change-Id: Id835b2263f4496525de21a0608d10a0dc6246e4c
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>