OMAP3+: PM: DVFS: Scale dependent domains to existing OPP
Currently dependent domain is scaled to:
- dependent voltage that is got from voltage dependency table.
- frequency that is got from OPP found for dependent voltage.

OPP for dependent voltage is found by _volt_to_opp() function
which looks for OPP with equal or _greater_ voltage.
So if there is no OPP with exact voltage (or this OPP is disabled),
then frequency from OPP with higher voltage will be used.

As a result device will run at not supported combination of frequency and voltage.
This may lead to instabilities.

This patch scales dependent domain to a correct voltage-frequency tuple,
by using both frequency and voltage from the matched OPP.

Signed-off-by: Taras Kondratiuk <taras@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>