gpu: ion: export needed functions for PVR SGX
Make ion symbols visible to kernel modules. The SGX DDK
in modular form requires these symbols to be exported.

The reason for the code movement of ion structs is due to
EXPORT_SYMBOL’s requirement that the storage size of all
function parameters is known. This is so that the kernel can
sanity-check the size of data structures at module load time.

Change-Id: I5797b228ec580fdd39fbf7f8e4d120c80b3d4409
Signed-off-by: Rodrigo Obregon <robregon@ti.com>