gpu: ion: export needed functions for PVR SGX
Make ion symbols visible to kernel modules. The SGX DDK
in modular form requires these symbols to be exported.

The reason for the code movement of ion structs is due to
EXPORT_SYMBOL’s requirement that the storage size of all
function parameters is known. This is so that the kernel can
sanity-check the size of data structures at module load time.

Ported to 3.8 ION by Sundar Raman<sunds@ti.com>

Change-Id: I52a4136ed47d83097c4c76e384ce11cd80ef3858
Signed-off-by: Rodrigo Obregon <robregon@ti.com>
Signed-off-by: Sundar Raman  <sunds@ti.com>