usb: dwc3: omap-glue: Extract dwc3 mem space from glue
The dwc3 glue holds the address space of both dwc3 wrapper(glue) and dwc3.
The first index holds the address of the wrapper and the second index holds
the address of the dwc3 core. The address specific to dwc3 is extracted
and used when creating a platform device for dwc3.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>