ASoC: twl6040: Let amics settle after biasing
After the analog mics are biased, they need a little time to settle.
However, the circuit is prevented from settling until the MicAmp is
enabled.  There is an also a 3ms pop when the ADC is enabled.

Since the settling noises are addative, the sleep to let them settle
is placed after the ADC enable because that is the last thing to be
enabled before beginning a capture.

The settling time is configurable because it is board-dependent, based
on the power circuit and the make/model of the microphone.

Change-Id: Ibd7e0bee4cd817c13575df6502d60a2f19527d9f
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>