usb: dwc3: core: add phy support
dwc3 core is now made to make use of the usb2 phy and usb3 phy
registered with the system. These phy's should be registered with otg
utility functions in order for dwc3 to use it. The phy's are switched on
when the module gets inserted and switched off when the module is
removed.

Change-Id: I65db408d8ec5ed962ab2b79f7952b47c4db99908
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>