OMAP: PM: register to the memory throughput class of the PM QoS framework
Implement the memory throughput constraints using the PM QoS notification
handler for the PM_QOS_MEMORY_THROUGHPUT class. The handler shall apply the
constraints aggregated value to the DVFS layer by calling the
corresponding function in the DVFS.

Note: The way requested throughput is converted to corresponding interconnect
frequency is still under discussion and hence this conversion mechanism needs
to be revisited.

Signed-off-by: Jean Pihet <j-pihet@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>