remoteproc: export rproc_error_reporter function
The rproc_error_reporter function is currently being used to
report MMU faults and trigger a recovery. This is internal to
the remoteproc framework.

However, the function can be used to report errors other than
just iommu faults, like errors that a remoteproc detects and
notifies the remoteproc driver. The rproc_error_reporter
function therefore is now exported to allow other clients to
report an error.

Change-Id: Ib4efd8a70e2ae745afbaaea5bca11f33ba07ad0a
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>