OMAPDSS: HDMI: Do not reset TV timings on sync loss
The panel needs to be disabled & enabled in case of SYNC_LOST.
However, HDMI panel clears EDID and the timings when this happens.
To avoid this, added a flag to notify device that we are handling a
sync loss.

Modified HDMI to not modify custom_timings and edid_set flags, so we
keep the resolution on power_on.

Change-Id: Idb3a776e057efff8efce677552e76c8b8de3053f
Signed-off-by: Baranwal Sujeet <s-baranwal@ti.com>
Signed-off-by: Muralidhar Dixit <murali.dixit@ti.com>
Signed-off-by: Lajos Molnar <lajos@ti.com>