OMAPDSS: HDMI: Moving HPD int handling to panel driver
Moving HPD handling to panel driver from IP file as current
HPD handling is not proper.

HPD interrupt is handled in ip specific file to handle the
phy burn out issue. On HDMI disconnect PHY is transitioned
to LDO_ON. This transition is triggered from HPD handler in
panel driver. On connect PHY is first moved to LDO_ON and
then after power on PHY is moved to TX_ON.
On HDMI enable PHY is first moved to LDO_ON and if HDMI is
connected then only PHY is moved to TX_ON.

Change-Id: Ieb6ccc6becce016859082e22b927cca8132ec18e
Signed-off-by: Muralidhar Dixit <murali.dixit@ti.com>