ASoC: abe mixer: Increase volume ramps for DMic's
Most digital microphones need about 10 ms to settle after setting bias
and clocks.  The OMAP DMic module is not able to set the clocks before
beginning the transfer, creating a full-scale pop at the beginning of
the transfer.

This patch works around this behavior by increasing the volume ramp
from 2ms to 50ms for DMic's.

Change-Id: I629dddc377a0d71770c0b812ec0b99933536fb2e
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>