HID: uhid: add UHID_CREATE and UHID_DESTROY events
UHID_CREATE and UHID_DESTROY are used to create and destroy a device on an
open uhid char-device. Internally, we allocate and register an HID device
with the HID core and immediately start the device. From now on events may
be received or sent to the device.

The UHID_CREATE event has a payload similar to the data used by
Bluetooth-HIDP when creating a new connection.

Change-Id: I4fcf07e1e580a90292a1ee9fbeb4d9124bf1d5eb
Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>