HID: uhid: add example program
This adds an example user-space program that emulates a 3 button mouse
with wheel. It detects keyboard presses and moves the mouse accordingly.

It register a fake HID device to feed the raw HID reports into the kernel.
In this example, you could use uinput to get the same result, but this
shows how to get the same behavior with uhid so you don't need HID parsers
in user-space.

Change-Id: I5002b5ad154dcf4cfdf45e3f9ea0c6e4500856b3
Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>