TMP006: Adding support for TMP006 IR temprature sensor
This patch provides the basic IR temperature sensor, TMP006, driver
to the Thermal framework.
This driver provides basic read of the current temperature. And it also
provides debugfs entry for current temperature measured. The sensor is
operated in default configuration of 1 cov/sec and in continuous conversion
mode.

debugfs location :
        /sys/kernel/debug/thermal_debug/devices/tmp006_sensor

Entries:
        current_temp    -- Current temp measured by IR sensor

Change-Id: I3cd21c19cb5c08c7d87678ac364877ae7f769345
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>