OMAP5:defconfig: Enable CRYTO_ECB required for SMP
SMP protocol part of Bluetooth Low Energy sub-system in kernel makes use of
this cryptographic libraries for its authentication and association with other
low energy devices (pairing).
SMP is dependent on ECB via Kconfig, However since Bluetooth sub-system is
built as modules, this library is also built as a module, This patch forces it
to be built as a part of kernel so that SMP can make use of it.

Change-Id: Ie89ffee0e87002cbbb1b249772e77c1e0ee96a7c
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>