OMAPDSS: APPLY: Add support for simultaneous multiple overlay updates
Description: The current apply mechanism in DSS enables/disables one overlay at a time, in order to have FIFO merge work. There is no provision of updating multiple
overlays in one shot, which is the most probable use-case with clinets like DSSCOMP and HLOS like Android.

In DSSCOMP, each composition may consist of multiple overlays, some of them
need enabling & some may need to be disabled. We need to update this entire info
in one single dispc_go, so all the layers get refreshed together.

This patch adds support for taking multiple overlays info & updating it
in one shot. The overlay manager gets the overlays info in an array that needs to
be updated together.

Change-Id: Ib9869f06720c8c6664dfc0e8b1a9b4d7379d3418
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>