rpmsg: resmgr: omap: allow _device_scale to call mach->ops function
The _device_scale function in the rpmsg omap resmgr code was stubbed
out previously as DVFS framework was not available in the kernel. This
would be used by resources such as IVA and FDIF to request a frequency
scale constraint. The DVFS code/framework implementation has been
added lately, and the stubbed code has therefore been cleaned up to
allow the _device_scale to actually call the .device_scale ops hook
registered through the platform data.

Change-Id: If8154a99fbcf4d7ddb200e0fae5215ea398912e2
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>