usb: dwc3: Add new gadget-ops to control pullup
Certain gadgets (like Android) require the pullup
to be low untill userland comes up, configures the
usb functions and then enables the pullup.

By default, the udc-core layer enables the pullup
during gadget_probe_driver, this causes break
in the cold boot enumeration on the Host for
such gadgets.

This patch introduces a new gadget-ops that
maintains a flag to allow/not-allow pullup
to happen if required by the gadget.

Change-Id: I12ecb58c70782e9be699edd8e825a3e38c2acb32
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>