ARM: OMAP2+: PM: ensure that platform device for L3 exists.
If DT is registered in the device, it is possible that DT definition
does not register OMAP L3 device in the system. Since it is expected
that when DT is used, ALL processor devices should have been provided
as well, not having processor device like L3 is never an acceptable
scenario.

Simply handling based on suggestion by Grygorii Strashko and use
omap_device_get_by_hwmod_name to lookup l3_device. BUG if we cannot
find the same so that platform bringup developer can fix it.

Change-Id: I79e1b7d5f7e1f25eca3a67a1217ebfee8fec1ba7
Signed-off-by: Nishanth Menon <nm@ti.com>