Filter touchscreen for erratic events
Logging suggests that a single touch can send events that vary
by as much as 70 points due to the highly sensitive touchscreen.
This patch attempts to capture the erratic events on the down
touch and not send the next touch event until it has moved
70 points (around 18 pixels).  The potential issue is loss
of sub 18 pixel move events after a down event.

Change-Id: I2a945a8ae4c5826d3be914e29f9a530eaa1580f4