ARM: OMAP2+: DVFS: remove req_dev from omap_device_scale
req_dev is used as an internal data structure inside DVFS layer.

Unfortunately, since this is exposed to the drivers, DVFS layer
must be capable of handling multiple request devices to the same
target device OR handle same request device request to multiple
target device.

DVFS layer is not geared up towards doing this and attempting to
do the same is debatable as to being able to cater to various
client user combinations which are possible. Further, a review
of usage of current and previous driver indicate that this
ability was never used in it's original purpose.

Hence, remove req_device as a parameter from omap_device_scale.

As part of this change, printing req_dev information has no
additional value, hence print the calling function in device
scale error logs to help track culprit caller down easily.

WARNING: This does break API between DVFS and driver users.

NOTES:
1) This patch removes ability as to requester to target device
mapping such as requests on pm_qos memory throughput tracking -
this needs to be provided as a debugfs support of some form
in a follow-on patch. With this patch, the dependency will
be reflected in <debugfs>/dvfs/core/info as l3 device instead
of the dummy device that was reflected previously.

2) The DVFS internal data structure still maintain a req_dev
tracking. But in reality this turns out to be:
 * All users of omap_device_scale req_dev == target_dev
 * For dependent domain scale requests req_dev = the device
   putting the constraint in.

3) If a need arises in the future where request to a specific
device needs an arbiter, a specialized API should be introduced
for that purpose.

As an existing example, on L3 device, OMAP implementation of
pm_qos memory throughput provides the necessary arbitration
required to determine the OPP that all the bus qos requests
map back to.

Change-Id: Ie3d91b1cf74d40e4cab2560fb5d60cd4f73fe60e
Signed-off-by: Nishanth Menon <nm@ti.com>