OMAP4 CPU Thermal governor:removes repetitive info
The thermal governor prints info after each crossing of a thermal treshold.
Often treshold is crossed but thermal zone was not changed. The printed info
to clogs a console and dosen't give an effective information. The patch
removes repetitive info for all zone except for the panic zone. The
panic zone is critical and a duplicate info must be.

Change-Id: If8bcbafe49a49a99f16dc362a3ac5d2f68251270
Signed-off-by: Eugen Mandrenko <ievgen.mandrenko@ti.com>
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>