omap5: bandgap: Enable IRQ towards the end of probe function
Enable the Interrupts once all the other parameters are set and
also any thermal framework related initialisation are done.
Otherwise irq handler might be called as soon as it is enabled
where as rest of settings/framework is still getting initialised.

Hence moved the IRQ initialisation towards the end of probe function.

Change-Id: I263aa844b59c061c1a92bd9a5cb1d64dc796ca2f
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>