Palmas: Fix power button to read the line status
The PWRON interrupt from palmas is only a falling edge condition so the driver will only
be called once when the button is pressed. And the current implementation reports a quick
press and release of the key to the input device driver.

To fix this there is a delayed work item that is scheduled when the
key is pressed to read the PWRON line status.

The work item will be scheduled @ 20ms intervals until the button is released.

Change-Id: I89ebcdec71372650e9478b32532da7a43a6c1c95
Signed-off-by: Dan Murphy <dmurphy@ti.com>