rpmsg: resmgr: omap: use pm qos api for memory throughput
The rpmsg resmgr driver expects the bandwidth API (for memory
throughput requests) to be defined by the device platform data.
This is no longer required since the pm_qos API would be used
for requesting memory throughput.

The patch removes the mach specific calls for throughput and
invokes the pm_qos APIs from the omap rpmsg resmgr driver for
all resources that support/require memory throughput on the
L3 bus.

Change-Id: I02c4ee51bdca99c62b02d0f2b7b5559098c63089
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>