remoteproc: omap: use pm qos api for memory throughput
The omap remoteproc driver expects the bandwidth API (for memory
throughput requests) to be defined by the device platform data.
This is no longer required since the pm_qos API would be used
for requesting the same.

The patch removes the machine specific calls for throughput and
calls the pm_qos APIs from the omap remoteproc driver.

Change-Id: Iad71a08017171c818915963963099a576af040f9
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>